Failure to read a mapping line results in no mapping
This is a weird bug on CentOS 7 where it fseek(SEEK_END) on a memory buffer doesn't
set EOF on the buffer, so the fgets() call still succeeds and returns an empty string.
sscanf() fails to read, but no_mapping doesn't get set to false so the test continues
to pass.